fix(flake): update code-server extensions

+++ flake.nix
@@ -380,13 +380,21 @@
                 # https://gist.github.com/hyperupcall/99e355405611be6c4e0a38b6e3e8aad0#file-settings-jsonc
                 installCodeServerExtensionsScript = pkgs.writeScript "install-code-extensions-run" ''
                   #!/command/with-contenv ${pkgs.runtimeShell}
+                  # disable "donjayamanne.python-environment-manager"
+                  # from "donjayamanne.python-extension-pack"
+                  # due to https://github.com/microsoft/vscode/issues/210792#issuecomment-2186965405
+                  # also leaves out
+                  # "wholroyd.jinja"
+                  # "batisteo.vscode-django"
                   VSCODE_EXTENSIONS=(
                     "alefragnani.project-manager"
                     "Catppuccin.catppuccin-vsc"
                     "charliermarsh.ruff"
                     "christian-kohler.path-intellisense"
                     "cweijan.vscode-database-client2"
-                    "donjayamanne.python-extension-pack"
+                    "ms-python.python"
+                    "njpwerner.autodocstring"
+                    "KevinRose.vsc-python-indent"
                     "eamodio.gitlens"
                     "github.vscode-github-actions"
                     "GitHub.vscode-pull-request-github"
@@ -395,7 +403,6 @@
                     "ms-azuretools.vscode-docker"
                     "ms-kubernetes-tools.vscode-kubernetes-tools"
                     "ms-toolsai.jupyter"
-                    "ms-vsliveshare.vsliveshare"
                     "njzy.stats-bar"
                     "patbenatar.advanced-new-file"
                     "rangav.vscode-thunder-client"
@@ -410,7 +417,8 @@
                   )
 
                   printf "Listing currently installed extensions...\n\n"
-                  code-server --list-extensions --show-versions
+                  installed_extensions=$(code-server --list-extensions --show-versions)
+                  echo "$installed_extensions"
                   echo ""
 
                   install_command="code-server"
@@ -420,10 +428,15 @@
 
                   eval "''${install_command} --force"
 
-                  tmpdir=$(mktemp -d) && \
-                  curl --proto '=https' --tlsv1.2 -sSfL -o "$tmpdir/vscode-just-syntax-0.3.0.vsix" https://github.com/nefrob/vscode-just/releases/download/0.3.0/vscode-just-syntax-0.3.0.vsix && \
-                  code-server --install-extension "$tmpdir/vscode-just-syntax-0.3.0.vsix" && \
-                  rm -r "$tmpdir"
+                  if echo "$installed_extensions" | ${pkgs.gnugrep}/bin/grep -q "nefrob.vscode-just-syntax"; then
+                      printf "vscode-just-syntax is already installed.\n"
+                  else
+                      printf "vscode-just-syntax is not installed. Proceeding with installation...\n"
+                      tmpdir=$(mktemp -d) && \
+                      curl --proto '=https' --tlsv1.2 -sSfL -o "$tmpdir/vscode-just-syntax-0.3.0.vsix" https://github.com/nefrob/vscode-just/releases/download/0.3.0/vscode-just-syntax-0.3.0.vsix && \
+                      code-server --install-extension "$tmpdir/vscode-just-syntax-0.3.0.vsix" && \
+                      rm -r "$tmpdir"
+                  fi
 
                   printf "Listing extensions after installation...\n\n"
                   code-server --list-extensions --show-versions
